feat: standardization code
