+++ package.json
@@ -11,6 +11,7 @@
   "author": "Axel <axeellima@gmail.com>",
   "license": "MIT",
   "devDependencies": {
+    "@types/cors": "^2.8.13",
     "@types/express": "^4.17.17",
     "@types/jsonwebtoken": "^9.0.1",
     "@types/node": "^18.13.0",
@@ -25,6 +26,7 @@
     "@types/uuid": "^9.0.0",
     "bcrypt": "^5.1.0",
     "bcryptjs": "^2.4.3",
+    "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.0",
