+++ src/@types/express/index.d.ts
@@ -5,7 +5,6 @@ declare global {
     interface Request {
       user: {
         id: string;
-        isAdm: boolean;
       };
     }
   }
