+++ src/app.ts
@@ -1,30 +1,21 @@
-import express, { response } from "express";
+import express from "express";
 import "express-async-errors";
 import "reflect-metadata";
-import { Request, Response, NextFunction } from "express";
-import { AppError } from "./errors/appError";
-import { userRoutes } from "./routes/user/user.routes";
+import { userRoutes } from "./routes/user.routes";
 import sessionRoutes from "./routes/session.routes";
-
-var cors = require("cors");
+import handleErrorMiddleware from "./middlewares/handleError.middleware";
+import cors from "cors";
+import { announcementRoutes } from "./routes/announcement.routes";
+import commentRoutes from "./routes/comment.routes";
 
 const app = express();
 app.use(cors());
 app.use(express.json());
 app.use("/users", userRoutes);
 app.use("/login", sessionRoutes);
+app.use("/annoucement", announcementRoutes);
+app.use("/comment", commentRoutes);
 
-app.use((err: Error, req: Request, res: Response, next: NextFunction) => {
-  if (err instanceof AppError) {
-    return response.status(err.statusCode).json({
-      status: "error",
-      message: err.message,
-    });
-  }
+app.use(handleErrorMiddleware);
 
-  return response.status(500).json({
-    status: "error",
-    message: "Internal server error",
-  });
-});
 export default app;
