+++ src/controllers/announcements/announcementDelete.controller.ts
@@ -1,11 +1,11 @@
 import { Request, Response } from "express";
 import announcementDeleteService from "../../services/announcements/announcementDelete.services";
 const userDeleteController = async (req: Request, res: Response) => {
-    const { id } = req.params;
+  const { id } = req.params;
 
-    await announcementDeleteService(id);
+  await announcementDeleteService(id);
 
-    return res.status(204).send();
+  return res.status(204).send();
 };
 
 export default userDeleteController;
