+++ src/controllers/announcements/announcementList.controller.ts
@@ -1,17 +1,10 @@
 import { Request, Response } from "express";
-import { AppError, handleError } from "../../errors/appError";
 import announcementListService from "../../services/announcements/announcementList.services";
 
 const ammountListController = async (req: Request, res: Response) => {
-  try {
-    const ammount = await announcementListService();
+  const ammount = await announcementListService();
 
-    return res.send(ammount);
-  } catch (error) {
-    if (error instanceof AppError) {
-      handleError(error, res);
-    }
-  }
+  return res.send(ammount);
 };
 
 export default ammountListController;
